requirements.txt
tensorflow
pandas
numpy
matplotlib
streamlit
Pillow
scikit-learn
